Added EditName and RemoveFeeder and made code more modular
This commit add both the EditName file as well as the RemoveFeeder file. As their names imply, these allow the user to edit the name of a feeder as well as remove it from their database. Also, I added a function to the database class that runs the usual checks I run for each file. This just reduces the amount of reused code and simplifies things.

Next, I need to handle the animal detected packet which will receive a packet with the devices MAC address so I can check the current time against that device's schedule. Depending on the result, a givefood packet will be sent back with either 0 cups of food or some amount. 0 means the animal wasn't allowed to eat and the red light will turn on. Otherwise, the amount of food will be given and a green light will turn on.