New Files to Handle Feeding Time List
As the title says, this commit adds two files for handling the feeding time list information. It also updates the other code to work with the newly named tables on the database.

The next thing I need to do is read the animal packet that will be posted by the Arduino, compare it against the time entry, and respond with a givefood packet with either no cups or the amount of cups shown in the  entry.