Added a PacketHandler file which isn't fully done
As the title says, I added a PacketHandler.php file. This will eventually contain all of the packets and their code. Doing this will consolidate the code quite a lot and simplify things (maybe). I'm not sure how much I want to follow through with this, but I will try it out.

I also modified a few lines of code to clean things up.